feat: merge gatsby static data hook
