+++ src/index.js
@@ -1,2 +1,1 @@
-export { usePrismicPreview } from './hooks'
-export { useNewPreview, usePreviewData } from './newHooks'
+export { useNewPreview, usePreviewData } from './hooks'
